Improve the precision of ledge detection
LedgeDetector now takes the initial offset of the rays in account.
Moved the rays to not be on the edges of the characters' collider: this would
trigger a bug that's hard to reproduce, where the character would jump back to
the origin of the world.

To get the player's look direction, the code now uses sign(_velocity.x)

The ledge state is still instant.