Close #40: remove level loops folder
The hook mechanic needs a redesign, and changes in the movement already makes some of the patterns obsolete. Also, moving to a tilemap. We need to iterate more on the design before saving reusable components like these. I think we should wait until we have graphics and create scenes that are more like more elaborate prefabs.