Hybrid Component/HFSM player skills system - 12.05.2019 Jam Report & PR
Hello there!

So although it still was a bit of time pressure (isn't it always?!), I
managed to come up with a satisfactory solution in my opinion - feedback
always welcome!

Upon research about HFSM, most implementations rely on inheritance. If a
substate can't handle an input then the event is moved up the parent
class. This works, but because of overlapping functionality it's not
always this clear cut so I went with a hybrid approach. It's not only
when the substate can't handle an event/input that it uses the
superstate, instead it uses the superstate functionality whenever
necessary.

In Godot, the inheritance approach isn't too good because internal
variables would be duplicated whereas we only need them in one place and
one place only. For example, the `Idle`, `Run`, `Air` substates which
are children of the `Move` state all update the `velocity` and it
wouldn't make sense for each of the named substates to have their own
instance `velocity` variable. Instead they all change the
`Move.velocity` variable directly.

Findings:

- it can be fairly complex to figure out how to structure the Skills
  tree: why would `Idle` be a child of `Move`? In our case it makes
  programmatic sense at least since it relies on some internal variables
  from its parent `Move` node, but not much more than that
- **only use leaf** nodes when changing state/skill. Most HFSM
  implementations force superstates to have an initial substate
  preselected. I chose not to do it this way as passing leaf nodes
  explicitly is clearer
- there's really no way to decouple entirely states/skills since they
  rely on information from each other. For example the `Ledge` skill
  needs velocity information from `Move`. With careful design it's
  possible to have the Godot scene Skills tree decoupled in the sense
  that if for example `Ledge` is missing then the player just looses the
  ability to be helped when near a ledge, but this needs some thought,
  it's not as simple as removing the `Ledge` node from the scene tree,
  some inter-dependent conditions need to be met in order for this to be
  safe

Conclusion:

With all the *downsides* from above, I think this system has potential
and it simplifies quite a bit the complexity, especially on the movement
implementation.

A number of improvements and simplifications have emerged as well:

- much simpler velocity calculation algorithm on simply both x & y
  components at the same time
- much simpler air/idle jump time delay implementation without the need
  of booleans to keep track of state: just use directly the timer's `time_left`
  property
- overall simplified code & decoupling

Known Issues:

- when using the hook mechanic at the moment the player detects a ledge,
  the skin animation fails to finish leaving the collision shape in an
  offset relative to the skin position

Notes:

I've changed the `LevelTest` scene to use the new `PlayerHFSM` scene,
while all the new states/skills go in `res://src/Player/SkillsHFSM` so
there's no overlap with previous code.

Hope it's clear enough - both explanation & code!