Core level load mechanic
I've fiddled with the existing nodes and used Level1 & Level2 as
test beds. Level2 was actually named Level3, but I renamed some
some nodes to keep consistency.

The idea is very simple, there's an Area2D (called Exit) that
behaves like a portal, which has an export variable that points
to the next level (on the disk) to be loaded.

When the Player is detected by the Area2D LevelLoader.trigger gets
called.

LevelLoader is now an autoloaded script which keeps track of the game
node, the player node and the current level.

When the game first starts the game node also uses LevelLoader to load
the start level, instead of having it present in the scene tree by
default.

note: I had to mess with the nodes like I said so let me know if there's
any conflicts I can help with.